+++ app/src/main/java/exp/ftxt/features/network_stats/NetworkConfig.java
@@ -20,4 +20,5 @@ public class NetworkConfig {
     public static int bgRadius = 0;
     public static float posX = 0.75f;
     public static float posY = 0.05f;
+    public static float updateInterval = 1f; // detik: 0.2, 0.5, 0.75, 1-10
 }